+++ src/main/java/com/javampire/openscad/psi/OpenSCADElementFactory.java
@@ -5,12 +5,26 @@ import com.intellij.openapi.project.Project;
 import com.intellij.psi.PsiElement;
 import com.intellij.psi.PsiFileFactory;
 import com.intellij.psi.tree.IElementType;
+import com.intellij.psi.util.PsiTreeUtil;
 import com.javampire.openscad.OpenSCADFileType;
 import com.javampire.openscad.psi.stub.function.OpenSCADFunctionStubElementType;
 import com.javampire.openscad.psi.stub.module.OpenSCADModuleStubElementType;
 import com.javampire.openscad.psi.stub.variable.OpenSCADVariableStubElementType;
 
 public class OpenSCADElementFactory {
+
+    public static PsiElement createImportPath(final Project project, final String path) {
+        final OpenSCADFile file = createFile(project, "include <" + path + ">;");
+        final OpenSCADIncludeImport include = PsiTreeUtil.findChildOfType(file, OpenSCADIncludeImport.class);
+        if (include == null) {
+            throw new IllegalStateException("Failed to create import path for: " + path);
+        }
+        final OpenSCADImportPathRef pathRef = include.getImportPathRef();
+        if (pathRef == null) {
+            throw new IllegalStateException("Failed to create import path for: " + path);
+        }
+        return pathRef;
+    }
 
     public static PsiElement createIdentifier(final Project project, final String name) {
         final OpenSCADFile file = createFile(project, name + "=0;");